Renamed bosch_hale.py to reactivity.py
